CORRECT: Изменил структуру проекта, а также добавил README.md и requirements.txt
