+++ requirements.txt
@@ -1,0 +1,6 @@
+uvicorn~=0.31.1
+fastapi~=0.115.2
+SQLAlchemy~=2.0.35
+requests~=2.32.3
+pydantic~=2.9.2
+databases~=0.9.0
